Restructing and creating dev upstream
