--- scripts/process_xml.py
@@ -1,1 +1,0 @@
-# TODO: Pre-rocess the XML file and extract the data and convert it into a pandas DataFrame
